Implement data_input.
Data_input is responsible for loading and processing the data.
Data input uses quese to do processing in the Background.
